REF: Move Tag and TagGroup into resources package
